Move org.jboss.as.test.integration.ejb.security.Util to shared module as per pskopek's request in AS7-2102.
