+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/security/WhoAmIServlet.java
@@ -36,7 +36,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.Writer;
 
-import static org.jboss.as.test.integration.ejb.security.Util.getCLMLoginContext;
+import static org.jboss.as.test.shared.integration.ejb.security.Util.getCLMLoginContext;
 
 /**
  * @author <a href="mailto:cdewolf@redhat.com">Carlo de Wolf</a>
