+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/security/base/EntryBean.java
@@ -27,7 +27,7 @@ import javax.ejb.SessionContext;
 import javax.security.auth.login.LoginContext;
 import javax.security.auth.login.LoginException;
 
-import static org.jboss.as.test.integration.ejb.security.Util.getCLMLoginContext;
+import static org.jboss.as.test.shared.integration.ejb.security.Util.getCLMLoginContext;
 
 import org.jboss.as.test.integration.ejb.security.WhoAmI;
