+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/security/rolelink/SecurityRoleLinkTestCase.java
@@ -28,7 +28,7 @@ import javax.security.auth.login.LoginContext;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.test.integration.ejb.security.SecurityTest;
-import org.jboss.as.test.integration.ejb.security.Util;
+import org.jboss.as.test.shared.integration.ejb.security.Util;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
