+++ testsuite/shared/src/main/java/org/jboss/as/test/shared/integration/ejb/security/Util.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.ejb.security;
+package org.jboss.as.test.shared.integration.ejb.security;
 
 import javax.naming.Context;
 import javax.naming.InitialContext;
